--- src/test/java/Model/SortFilter/TestColumnData.java
@@ -1,32 +1,0 @@
-package Model.SortFilter;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-class TestColumnData {
-
-  @BeforeEach
-  void setUp() {
-  }
-
-  @Test
-  void getColumnName() {
-  }
-
-  @Test
-  void fromColumnName() {
-  }
-
-  @Test
-  void fromString() {
-  }
-
-  @Test
-  void values() {
-  }
-
-  @Test
-  void valueOf() {
-  }
-}
