+++ src/test/java/TestController.java
@@ -1,5 +1,3 @@
-package Controller;
-
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
