+++ src/test/java/TestJFrameView.java
@@ -1,5 +1,3 @@
-package View;
-
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
