+++ src/test/java/TestOperations.java
@@ -1,5 +1,3 @@
-package Model.SortFilter;
-
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
