+++ src/test/java/TestPlayer.java
@@ -1,9 +1,12 @@
-package Model;
-
 import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class TestPlayerBean {
+class TestPlayer {
+
+  @BeforeEach
+  void setUp() {
+  }
 
   @Test
   void getName() {
@@ -43,10 +46,6 @@ class TestPlayerBean {
 
   @Test
   void getPpg() {
-  }
-
-  @Test
-  void getRpg() {
   }
 
   @Test
