+++ src/test/java/TestPlayerBean.java
@@ -1,14 +1,7 @@
-package Model;
-
 import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class TestPlayer {
-
-  @BeforeEach
-  void setUp() {
-  }
+class TestPlayerBean {
 
   @Test
   void getName() {
@@ -48,6 +41,10 @@ class TestPlayer {
 
   @Test
   void getPpg() {
+  }
+
+  @Test
+  void getRpg() {
   }
 
   @Test
