+++ src/test/java/TestPlayerSort.java
@@ -1,0 +1,18 @@
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class TestPlayerSort {
+
+  @BeforeEach
+  void setUp() {
+  }
+
+  @Test
+  void sortPlayers() {
+  }
+
+  @Test
+  void testSortPlayers() {
+  }
+}
